backend/requirements.txt
# Django Core
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg[binary]==3.2.13

# LLM Provider
anthropic==0.40.0

# Async / Queue
celery==5.4.0
redis==5.0.8

# Utilities
python-dotenv==1.0.1
django-extensions==3.2.3
jinja2==3.1.3

# Development
black==24.3.0
pytest==8.0.0
pytest-django==4.8.0
